// apps/marginfi-v2-trading/src/components/common/Pool/create-pool-dialog/types.ts
import { BankConfigOpt, OracleConfigOpt } from "@mrgnlabs/marginfi-client-v2";
import { PublicKey } from "@solana/web3.js";

export enum CreatePoolState {
  TOKEN = "token",
  QUOTE = "quote",
  FORM = "form",
  CONFIGURE = "configure",
  REVIEW = "review",
  LOADING = "loading",
  SUCCESS = "success",
  ERROR = "error",
}

export type PoolMintData = {
  mint: PublicKey;
  name: string;
  symbol: string;
  icon: string;
  decimals: number;
  price: number;
};

export type PoolData = {
  token: PoolMintData;
  quoteToken?: PoolMintData;
  tokenConfig?: { bankConfig: BankConfigOpt; oracleConfig: OracleConfigOpt | null };
  quoteTokenConfig?: { bankConfig: BankConfigOpt; oracleConfig: OracleConfigOpt | null };
  group?: PublicKey;
};